Simplify sidebar by splitting into components Reuse workspace search component into Workspace card on dashboard allow customized/override classes for display of workspace items
